+++ src/modules/fancyzones/editor/FancyZonesEditor/MonitorVM/Command.cs
